move handle utils
