+++ packages/sdk/src/query/utils.ts
@@ -60,16 +60,6 @@ export type StrippedQueryOptionKeys =
   | "query"
   | "pollingInterval";
 
-export const ZERO_HANDLE =
-  "0x0000000000000000000000000000000000000000000000000000000000000000" as const;
-
-/**
- * Check whether a handle represents the zero value.
- */
-export function isZeroHandle(handle: string): boolean {
-  return handle === ZERO_HANDLE || handle === "0x";
-}
-
 /**
  * Remove TanStack behavioral options from a query config object so only domain
  * parameters remain for the lower-level factory.
